i#4865 emulate: Use new API in drmemtrace
Uses the new emulation instrumentation interface in drcachesim's
drmemtrace tracer.  Removes all of the special-case code for both
online and offline tracing that is no longer needed now that a general
emulation approach is used for rep string and scatter expansion.

Confirmed that the allasm_repstr and allasm_scattergather applications
have the exact same behavior after the changes.

Fixes #4865