i#4719 qemu: Replace WFI with WFI for QEMU interop
QEMU crashes when executing the WFI AArchXX instruction.
DR uses WFI in its spinlock loops.
We replace WFI with WFE here, which is roughly a superset
of WFI and should work similarly for us while not
breaking QEMU.

This fixes over a dozen tests that previously failed under QEMU with
"unhandled CPU exception 0x10001".  Here we add 17 tests on AArch64
and 14 tests on AArch32 to the RUNS_ON_QEMU label.

Issue: #4719