Fix quoting mishap in build-from-source docs
Apparently Doxygen's markdown parser cannot handle a single backtic
inside a triple-backtic quote block.  This was causing a messed-up
sequence of what is quoted and what is not and a missed final section
on the build-from-source web page.  We fix it here by switching to $()
instead of backtics in a bash function; we also add a named link to
the DrGUI build directions since Dr. Memory links to them.