i#4274 kernel pc: Export version number in final trace
The version number in a drmemtrace trace is currently hidden in a
header that the reader hides from analysis tools.  Here we add a new
marker type that exports the version number so that analysis tools can
handle legacy traces.

Adds version awareness to the view tool.

Tested by manually running the view tool:
  $ bin64/drrun -t drcachesim -verbose 3 -- suite/tests/bin/allasm_repstr
  ::4152878.4152878:: marker type 12 value 2
  ::4152878.4152878:: marker type 9 value 64
  ::4152878.4152878:: marker type 10 value 64
  ...
  $ bin64/drrun -t drcachesim -offline -- suite/tests/bin/allasm_repstr
  $ bin64/drrun -t drcachesim -simulator_type view -indir drmemtrace.allasm_repstr.*.dir
  <marker: version 2>
  <marker: filetype 40>
  <marker: cache line size 64>
  <marker: timestamp 13269494545173426>
  <marker: tid 4152931 on core 2>
    0x0000000000401000  48 83 e4 f0          and    $0xf0, %rsp
  ...

Issue: #4274