i#4963 drreg state restore: handle non-drreg spills.
Revamps drreg's state restoration logic (again) to not need to observe non-drreg
spills and restores.

The existing strategy of tracking the app value across spills and restores
required drreg to be aware of non-drreg spilling mechanisms like spilling to
mcontext, to stack, or any other possible methods. This makes it non-ideal.

Implementes a new strategy for state restoration that does not need to be aware
of non-drreg spills, and also handles the various tricky drreg cases like
multi-phase use. The key observation behind this is that it is easier to find
the matching spill for a given restore, than to find the matching restore for a
given spill. This is because there may be other restores besides the final
restore (e.g. restores for app read, user prompted restores, etc.). This makes
it hard to find exactly where the spill region for a reg/aflags ends. Now we
walk the ilist backwards from the last instr to the faulting instr. The first
restore (or last restore in the list) found is expected to be an app value
restore, we find the matching spill for that restore which uses the same slot
which has to be an app value spill too. We continue doing this until we reach
the faulting instr and if there's a spill-restore pair where the spill is before
the faulting instr, we use that slot to restore the reg/aflags.

Adds multiple new test cases that use non-drreg spilling mechanisms: spilling
to mcontext, clean call instrumentation. Also modifies some previous tests so
that the interesting part of the test happens after the faulting instr; this is
required because we now walk backwards in the ilist, not forwards.

Issue: #3823
Fixes: #4963