i#4998: Fix incorrectly nested signals (#5008)
Fixes a bug where DR delivers a signal while an app is inside its
signal handler for that same signal and has blocked such self-nesting.
The fix is to undo some of the sigsuspend-handling changes that cause
this regression by moving the sigsuspend state restore to actual
delivery rather than at pending signal time.  The ignoring of the current
blocked signal set does have to remain in some form to handle
pselect and signals like it where we can't move the post-syscall
action to after the signal delivery.

Improves the linux.signest test to reproduce the original bug and
verify the fix.  Also adds a test case to ensure that signals are
nested when not blocked.

Fixes #4998