i#5009: Functionality for counting compulsory misses
This patch contains implementation of the functionality for counting
compulsory misses as a separate metric.

In this implementation information about all the already accessed memory
regions are stored. If the cache miss happens and block which given address
refers to was never accessed we are counting it as a compulsory miss.

Issue: #5009