i#5076 trace invariants, part 3: Add simulator_type
Adds a new -simulator_type invariant_checker to support a user
launching the checker explicitly on a target trace.

Adds documentation on the new top-level tool.

Adds a simple sanity test of the new flag.  (The -test_mode versions
of existing tests are already stressing the actual invariant checks.)

Issue: #5076