i#5070: AArch64 v8.0 decode: Fix issues with instruction differentiation
The AArch64 codec currently uses masking to
determine the instruction and where its operands
lie. This is somewhat limiting because AArch64 can
have instructions that appear to overlap using this
method.

This patch adds the + character to operands which
requires that at least one of the bits marked + must
be set for that instruction to be considered matching.

In addition, this patch switches the decoder to generate
else if statements rather than assuming that every branch
contains a return.

Issues: #5070
Change-Id: Iaf734887e4c45d658db68b5286fa661464238504