i#4924: Enable drcachesim -test_mode in release builds
Makes the -test_mode option, which runs the invariant_checker tool,
available in release builds.  This fixes the drcacheoff.rseq test
failures on GA CI when rseq is actually available in our upgrade to
Ubuntu 20.04.

Removes the test failure ignore rule for Jenkins as this also fixes
the AArch64 failure for the same test.

Fixes #4924
Issue: #4924, #4953