i#5079: Avoid skipping get_pc_thunk after redirect.
Moves accesses to global variables and string constants to a separate method
that we always execute from the beginning.

Jumping to the middle of a routine using DR_SIGNAL_REDIRECT causes us to skip the
__x86.get_pc_thunk.x calls that are required before accesses to global variables
in PIC. This call allows accessing them using a fixed offset from code, by
loading the pc of code in a register. Skipping this call may cause the program
to crash due to a bad address.

This fixes the client.signal test on 32-bit, which would otherwise keep crashing
in an infinite loop (the crash loop is due to other setup in this test).

Issue: #5079