i#4953: Upgrade x86-32 CI to use Ubuntu 18.04 (#5072)
Upgrades Ubuntu version for the x86 32-bit CI job to 18.04. We still have some
tests failing on that -- tool.histogram.offline and common.decode (flaky); added
them to ignore list for now as we want to upgrade ASAP before Ubuntu 16.04
jobs start failing on Github Actions.

The linux.clone-reset failure is a different existing issue. #4604

Other CI jobs have been updated to 20.04 already. But x86-32 has some
more failures on 20.04; we use 18.04 for now so that we have to ignore
lesser number of CI failure:
https://github.com/DynamoRIO/dynamorio/pull/5108#issuecomment-922418914

Issue: #4953