i#4678: AArch64 fuzz app for drstatecmp.
Adds a fuzzing app for AArch64 that automatically generates from the decoding files a random sequence of side-effect-free and non-branch instructions with fuzzed operands for stress-testing clients with drstatecmp. This enables automatic testing of even newly inserted instructions in the decoder.

Adds a drcachesim test with delayed tracing for the fuzzing app. Testing with an older DR version showed that this test would have eventually automatically detected #5023 (instrumentation-clobbering of aflags).

Issue: #4678