i#5121: Add test for getting app value in multi-phase drreg
Adds a test that demonstrates how to get the app value for a reg
that was reserved in multiple phases using drreg. This is complex
as in the current design a phase doesn't know about spill done
in earlier phases. So, getting the app value in insertion phase
for a reg that was reserved in app2app and insertion phases
both requires both the phases to work together.

Adds to the documentation of drreg_get_app_value and
drreg_statelessly_restore_app_value that they allow obtaining
the reg's meta value from the last phase where it was reserved,
OR the app value if there's no such previous phase.

Issue: #5121