i#5129 Fix client_ibl_xfer to avoid overwriting next block
Fixes the case on AArch64 where the LDR+BR jump instructions in the
client_ibl_xfer block can overwrite the first instruction of the next
generated block. This happens when relink_special_ibl_xfer() is called
AND the cache line alignment places the LDR+BR at the end of
client_ibl_xfer.

Issue: #5129