i#5131 new linux syscalls: Handle clone3
Adds  support for the clone3 syscall, similar to existing support  for the clone
syscall.  The functionality provided by  clone3 is a superset of clone.  It uses
a struct to encapsulate all args.  Also, the stack address is expected to be the
lowest one in  the stack (inclusive);  this is unlike clone where it is expected
to be the highest one (non-inclusive).

Extends the existing linux.clone test to verify behavior on clone3 as well.