refactor install of cmake-generated files
This patch replaces the install(CODE) commands that generate
files on the fly.
Instead, the files are generated at configuration time and just copied
at install time. By that, install relocation via various strategies like
DESTDIR and CPack works correctly. Further, this gets rid of a lot of
complex escaping and makes the results more predictable.

When appending to cmake-generated files, we also ensure that the file
is cleared on a re-configuration run to avoid duplicates in the file.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>