i#5131 new linux syscalls: handle pselect6_time64
Adds handling for pselect6_time64 same as the existing handling for
pselect6.

Also adds a test for pselect6_time64 with non-null sigmask that
fails without this. The one with null sigmask doesn't really need
special handling but added it anyway.

Also adds a test for ppoll_time64 with non-null sigmask that fails
without the ppoll_time64 handling. The existing one for ppoll_time64
was with null sigmask. The existing non-null sigmask test for ppoll
also does fail without the ppoll_time64 handling, but good to have
another one specific to ppoll_time64.

Issue: #5131