debian: add initial debianization for testing GNU style install
This patch adds an example debianization to test the
correct functionality of the GNU style install tree.

To simplify execution from system-wide installs, a wrapper
"drrun-system" is added around drrun to point the DR root
to "/".

Note: This patch should not be integrated into DynamoRIO.
Instead it should be proposed for Debian once the other
patches are integrated into DynamoRIO.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>