add support for GNU style directory layout
This patch adds a cmake switch USE_GNU_DIR_LAYOUT to change the
layout of the install tree. By that, the previous install tree
layout remains the default as log as the option is not activated.

The GNU layout uses the cmake-builtin GNUInstallDirs package
to layout the directories according to the GNU coding standards.
This includes multi-arch support on Debian as well.

Customizations of the layout are possible by overwriting the corresponding
cmake variables like CMAKE_INSTALL_<...>.

Further, all libraries (including clients) are built with soversions.
This ensures that multiple versions of DynamoRIO libraries can be
installed on the same system at the same time.
This is propagated to both internal and external clients as well (via the
DynamoRIOConfig cmake script).

The DynamoRIOConfig script is extended to auto-detect and support both
the traditional and the GNU layout, so that external clients
can be integrated as previously.

Note that this patch also unifies directory names by using the
corresponding variables instead of hard-coded path across
all CMake scripts.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>