RFC: drrun: add support to inject from a gnu-dir-layout DR install
This patch extends the DynamoRIO frontend and drdeploy to
support injection from a system-wide install with GNU layout.

To activate the support, pass "/" as DynamoRIO root.
This should not conflict with traditional DR roots, as a traditonal
install into the system's root folder is not likely or expected.

The new injection logic works (at least on x86) targets for both
release / debug and i386 / x64 versions.

Further, the drrun from a GNU layout install can be used to execute
from a traditional DynamoRIO root by just specifying the path
to the directory.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>