Relax the just-added assert on clean call flags during insertion phase only: we have to allow DR to invoke the callback during mangling.
Fix drreg-test cases that now have failures coming from the bug fix in
restoring xax when it's dead.

drreg-test has cases like Test 21 where xax is dead but it does this:

    /* Make sure that aflags are spilled to some slot, instead of being stored in xax.
     */
    CHECK(drreg_get_app_value(drcontext, bb, inst, DR_REG_XAX, DR_REG_XCX) ==
              DRREG_SUCCESS,
          "cannot get app value");

That now fails, which it should do since xax is dead and so we can't
actually get the app value.

I changed the 7 drreg test asm sequence to have xax be live, so this
sort of hack to get the flags out of xax still works.