Fix drdisas handling of leading zero bytes
On AArch64, the drdisas input can have zero bytes which are not
correctly handled.  We fix that here by appending the proper number of
zero bytes rather than always just one.

The simple test is augmented to include a fully-zero encoding, which
trips up the entire sequence without this fix.