Fix crash due to bad redirect in some env.
Fixes a crash due to jumping in the middle of a routine in
the client.signal test. We disable inlining so that we
execute the saving/setting of the frame pointer after
starting execution from the redirect pc.

Starting execution from the  middle of a routine can cause
issues if essential code sequences like this are skipped.