i#5076 trace invariants, part 6: Count memrefs in view tool
Changes the view tool's -skip_refs and -sim_refs behavior to count
memrefs rather than instructions, for consistency with other tools.
It was only counting instructions at all because until recently when
it was augmented to print data and markers it only printed
instructions.

Removes the feature of printing the version and filetype for
-skip_refs as the usefulness is outweighed by messing up the reference
count (not an issue with the instruction count).

Adds a new unit test for the view tool which mocks the module mapping
and tests the skip_refs and num_refs options.

Issue: #5076