i#5131 new Linux syscalls: fix clone3 struct handling
Fixes DR's handling of the clone_args struct used by the app. While creating our
own copy of the app's clone_args object, we need to allocate as much space as
specified by the app in the clone3 syscall's args. We should not use the size of
DR's internal struct copy (clone3_syscall_args_t) for this.

Adds comments to explain this in various places.

Issue: #5131