i#5190: drcachesim child stats include all lower levels
The child (==lower level) cache stats in drcachesim today are only for
the immediately lower level. This is a holdover from the original
implementation which didn't support other than 2 levels. For multiple
levels, one would expect the child stats to include the sum of all the
lower levels.  We fix that here and add a unit test.

The difference is clear in the config file test:
  LLC stats:
     Child hits:                        147
=>
  LLC stats:
     Child hits:                    174,695

But to avoid flakiness we do not update its regex and rely on the unit
test.

Fixes #5190