i#5131 new linux syscalls: handle close_range.
Adds handling for the close_range syscall in Linux. close_range allows either
closing a contiguous range of FDs or making them as close-on-exec. We convert
the seen close_range call into multiple such calls, one for each contiguous
sub-range that does not have any DR-private FDs. We always skip executing the
app's close_range.

Fixes the file_io test that assumed that there are only 4096 FDs. Instead, we
get the rlim_max for RLIMIT_NOFILE and try to use the DR_STEAL_FDS FDs after it.

Adds tests for close_range to verify correct happy-case behaviour and EINVAL.
close_range does not propagate errors from closing individual FDs; tests that as
well.

Manually defines the SYS_close_range constant to __NR_close_range. This is
needed because the header using which we updated the DR syscall table defines
the latter but not the former.

Issue: #5131