i#5131 new linux syscalls: Fix aarch64 dup3 and close_range handling
Fixes an existing bug in dup2/dup3 handling on aarch64. As SYS_dup2 is not
defined on aarch64, SYS_dup3 was not being included in the switch in
pre-syscall handling due to a misplaced conditional compile statement.

Fixes close_range handling on aarch64 by updating the syscall defines from a
newer header file that does define __NR_close_range.

Enables the client.file_io test on aarchxx. Replaced the inline asm code to
unmask divide-by-zero exceptions with a glibc wrapper.

On AArch64, unmasking of float-point exceptions may be ignored as trapping
exceptions is optional.
https://code.woboq.org/userspace/glibc/sysdeps/aarch64/fpu/feenablxcpt.c.html#37

Issue: #5131