i#2985 scatter-gather: Fixes handling of scratch xmm reg.
Fixes spill and restore for scratch xmm reg. Writing to the lower
xmm bytes zeroes the upper bits of the corresponding ymm and zmm
reg. So we need to spill and restore the entire ymm or zmm reg,
whichever is the largest mm reg supported by the system.

Also replaces the linear address of spill slot with an address
computed online using the TLS reg.

Issue: #2985