i#2985 scatter-gather: Fix spill/restore of scratch xmm reg. (#5252)
Fixes spill and restore for the scratch xmm reg. Mov-ing to the lower
xmm bits zeroes the upper bits of the corresponding ymm and zmm
reg. So we need to spill and restore the entire ymm or zmm reg,
whichever is the largest one reg supported by the system, and not
just the xmm part of it.

Also adjusts the state restoration machine to expect the new spill
instrs.

Replaces the linear address of spill slot in instrumentation with an
address computed online using the TLS reg.

Extends the allasm-scattergather test to verify the whole zmm reg
after the scatter/gather expansion. On systems without AVX512, we
verify ymm. This test fails without the fix.

Issue: #2985